ai-trader-bot/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
ccxt
requests
pydantic
